separate folders for the routes + initla setup for login wiht middleware
