+++ routes/auth_routes.py
@@ -1,0 +1,61 @@
+# routes/auth_routes.py
+from fastapi import APIRouter, Depends, HTTPException, status, Request, Response
+from fastapi.security import OAuth2PasswordRequestForm
+from fastapi.responses import RedirectResponse
+from sqlalchemy.orm import Session
+from datetime import timedelta
+from fastapi.templating import Jinja2Templates
+
+from auth import (
+    verify_password,
+    create_access_token,
+    ACCESS_TOKEN_EXPIRE_MINUTES,
+    get_current_user
+)
+from models import User
+from base import SessionLocal
+
+router = APIRouter()
+templates = Jinja2Templates(directory="templates")
+
+@router.get("/login")
+async def login_form(request: Request):
+    return templates.TemplateResponse("auth/login.html", {"request": request})
+
+@router.post("/token")
+async def login(
+    response: Response,
+    form_data: OAuth2PasswordRequestForm = Depends(),
+    db: Session = Depends(SessionLocal)
+):
+    user = db.query(User).filter(User.username == form_data.username).first()
+    if not user or not verify_password(form_data.password, user.password):
+        raise HTTPException(
+            status_code=status.HTTP_401_UNAUTHORIZED,
+            detail="Incorrect username or password",
+            headers={"WWW-Authenticate": "Bearer"},
+        )
+    
+    access_token_expires = timedelta(minutes=ACCESS_TOKEN_EXPIRE_MINUTES)
+    access_token = create_access_token(
+        data={"sub": user.username}, expires_delta=access_token_expires
+    )
+    
+    # Set cookie with token
+    response.set_cookie(
+        key="access_token",
+        value=access_token,
+        httponly=True,
+        max_age=1800,
+        expires=1800,
+    )
+    return {"access_token": access_token, "token_type": "bearer"}
+
+@router.get("/logout")
+async def logout(response: Response):
+    response.delete_cookie("access_token")
+    return RedirectResponse(url="/login")
+
+@router.get("/profile")
+async def read_users_me(current_user: User = Depends(get_current_user)):
+    return current_user
